requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
Faker>=18.0.0
reportlab>=4.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
